powerpc: Added initial device tree for MAIO400 board
powerpc: maio400: Added support for the onboard 16MB SPI flash.

powerpc: maio400: Added support for LED D55 (the green LED)

powerpc: maio400: Added support for LEDs D16 & D17.

These are the LEDs that you can see on the edge connector.
D16 is the top yellow LED.
D17 is the bottom green LED.

powerpc: maio400: Partitioned the SPI Flash

powerpc: maio400: Move and shrink the PCI windows.

Due to the previous changes to memory layout we
need to shrink and move te PCI windows.

Revert "powerpc: maio400: Move and shrink the PCI windows."

This reverts commit f9f0d5bbaa996e1c2107bf713f4ac9fd4e3cefc0. The commit
breaks at least the maio FPGA drivers - the drivers can be loaded, but the
test programs hang or do not work any more.

powerpc: maio400: Correct the pci addresses in the device tree (again).

maio400: Add support for an external clock.

powerpc: maio400: Fixed switched fiper1 and fiper2 values

powerpc: maio: Add an entry for the ICS9FG104 clock generator chip

This tree node communicates information about the ICS9FG104 clock generator
chip and the used interrupt lines to the maio-ck-req-mpcie driver.  This
driver is used to control the mPICE reference clock lines.

powerpc: maio: Add an entry for the Power-Fail interrupt.

This tree node tells the maio-pwr-fail driver which interrupt to look at.
This interrupt line is toggeled when the power is about to fail.

powerpc: maio: Clean up the device tree a little.

Remove some incorrect interrupt lines.  Our mdio chips are not assigned
interrupt lines.  Also remove a stale comment block.

powerpc: maio400: Switch the assignements of PCI interrupts

We cannot share an interrupt line between an external interrupt
source and a mapped PCIe interrupt.  As the power-fail interrupt
is on IRQ-0 we cannot use IRQ-0 for a PCI interrupt.

The solution is to switch the assignments of the interrupts.

This is the best reference that I have found.
http://www.devicetree.org/Device_Tree_Usage#Advanced_Interrupt_Mapping

powerpc: maio: Add FPGA and remove debug LED

The MAIO-400-2 board programs the FPGA via the CPU, so we should
add and entry to the device tree for this SPI device.  GPIO 12
is now part of this programming circuit and is no longer useable
as a debug LED, so remove it.

Technically these changes should only be applied to verion 400-2
boards, but they will have no real effect on 400-1 boards (of which
there are only a handful).

powerpc: maio400: Correct bug in dts file

powerpc: maio400: Reduce the SPI bus speed to 25 MHz

p2020maio401.dts: switch to omicron,maio400 machine