maio400: Add a MAIO-400 machine
The MAIO-400 machine is based on the code in mpc85xx_rdb.c and
mpc85xx_ds.c  The difference is that for the MAIO-400 we attempt
to enter a very low power mode after power_off.  This is important
because one of the reasons that we may have powered off was due to
an over-temperature condition, and it is important to try to lower
the temperature as quickly as possible.

The swiotlb support is required for the maio pcie driver - the FPGA
cannot access memory above 1 Gig, so we need to enable SWIOTLB if
needed.

Signed-off-by: Mark Marshall <mark.marshall@omicron.at>