mic400: Add a MIC-400 machine
Based on p1010rdb.c.  The main difference is that we atempt to
enter a low power mode on power off.

Signed-off-by: Mark Marshall <mark.marshall@omicron.at>

mic400: Add a restart handler (that does nothing)

It is not possible to reset the MIC-400 from the CPU, so add a
handler that does a halt instead.

Signed-off-by: Mark Marshall <mark.marshall@omicron.at>

mic400: Remove code to disable board reset

We will soon have MIC400 boards that have HRESET_REQ wired!

Added a P1010 reset function.

This function sets the "SW_RR" bit in the RSTRSCR register of the
GUT. This pulls the HRESET_REQ ouput line to active and the reset logic should
then  activate the HRESET input line so that the SoC resets.

mic400: Remove custom reset handler

(and damage done by commit 64f8c6220177c274f1d2ceac9a38d589cbd2d28a /
 37cc8fd095ba9e30fe7c05058c45192e3f9c3aaa)

Add another method for resetting a P1010

This function also does not work. Leave it in here until the HRESET_REQ
and HRESET line problem is sorted out.

Conflicts:
	arch/powerpc/platforms/85xx/Kconfig

mic400: Do not switch off the E500 or TimeBase when we halt

This has the effect of keeping the watchdog running when we halt.
This means that we will then power back on after a shutdown.