xhci: Add a patch to xhci to download firmware (from Renesas)
The Renesas USB 3.0 chip used on the MAIO board needs some firmware
downloading to it after power up.  This is normally done either by
the BIOS or with an external EEPROM.  As we have neither we need to
download the firmware at some point.  Doing it in the kernel means that
we can upgrade the firmware image.