gianfar: Work around race in gfar_poll_rx
There's a race between checking the ring (in gfar_clean_rx_ring) and
clearing the rstat:RXFn bit.  If the ring gets new packets between
the clean_rx_ring call and the write above, we will stop the RX path.

The solution is to check the ring again, if there are any new packets
in it we don't reduce the number of active queues.  We also have to
save the fact that the rstat:RXFn bit is now (possibly) lying to us.

This bug was only exposed after the changes to the RX buffer management
were made.  We now allocate a bunch of buffers at the end of
clean_rx_ring, which makes the "race window" much bigger.