gianfar: Optimize filer by removing repeated, identical, set-mask commands
A common command in the filer is the one to set the mask.  If the mask is
the same as the previous set-mask command, it can be skipped.  Often the
mask is the same (~0).

The only additional complexity is that we have to be care when we add
the CLE of the start of a group, as that might start at old_index or
old_index + 1.