gianfar: Split the RX processing of packets between two threads
We can now set a bitmask of which Q's are NAPI and which are "real-time".
The bitmask is in IBM order (Q0 = 0x80, Q1 = 0x40, Q7 = 0x01).  This
means that if the mask contains a 1, that queue is processed in the
normal way, via a softirq handler running the NAPI code.  If the mask
contains a 0, then that queue is handled directly in the threaded
interrupt handler.

We also modify the call to request_irq to disallow the processing of soft
irq's from within the hard receive irq.  This is required so that the
NAPI Q's are processed with a lower priority.

I have reversed the order of rx_bit_map and tx_bit_map so that they
now match the order of the other bitmaps in the driver.  This allows
me to "and" the bitmaps together.