gianfar: flow-classification: Allow to match against the first packet bytes
If we setup a rule with NFC, and the extension flow section is set, we can
match on an arbitrary four bytes from the packet.  If the second 32-bits
of data are written to RBIFX.  If this is 0x80818485 then we match against
bytes 0,1,4,5 of the packet (following the layer2 header).  These bytes are
the AppID and Flags from a Sampled Values packet.  The first 32-bites
of the extended data gives the values to expect in those bytes:

ethtool --config-nfc eth0 flow-type ether \
	proto 0x88BA action 5 \
	user-def 0x6d6e800080818485 m 0x00007fff00000000 loc 130

The above line will match only packets with an APP-ID of 0x6d6e and with
the Simulate flag set to 1.

This extension is a little non-standard!