gianfar: Use the napi RX budget more fully and more fairly
We now use much more of the budget than before, and try to share it out
among the different Q's more evenly.  With the previous code, if we had
a budget of 64, and Q depths of 2, 8 & 100, we would have only processed
the first 2 Q's fully and 21 items from the third Q, when NAPI is
expecting us to process 64 packets (54 from the third Q).  We will now
do that.

Clear RHALT even when Q is not empty.