tpm_tis_spi: Add an "ignore-wait" option to the TPM TIS SPI driver
This option indicates to the driver that it should ignore the "wait state"
feature of the TPM TIS spec.  This is designed for SPI controllers that
cannot split up the SPI transaction while keeping the CS line asserted
(such as the fsl-espi driver, "fsl,mpc8536-espi").

This mode is breaking the TPM spec!!